requirements.txt
alipay==0.7.4
amqp==1.4.9
anyjson==0.3.3
astroid==2.2.5
billiard==3.3.0.23
celery==3.1.25
certifi==2018.4.16
chardet==3.0.4
colorama==0.4.1
comtypes==1.1.4
demjson==2.2.4
Django==2.0.7
django-haystack==2.8.1
django-redis==4.10.0
django-tinymce==2.8.0
fdfs-client-py==1.2.6
itsdangerous==1.1.0
jieba==0.39
lazy-object-proxy==1.4.1
Pillow==5.2.0
pycryptodomex==3.7.2
pylint==2.3.1
pymysql==0.8.1
pypng==0.0.18
PyQRCode==1.2.1
python-alipay-sdk==1.10.1
python-devtools==2
pytz==2018.5
pywifi==1.1.8
redis==2.10.6
regex==2018.7.11
requests==2.19.1
urllib3==1.23
vine==1.3.0
virtualenv==16.6.1
virtualenvwrapper-win==1.2.5
Whoosh==2.7.4
wrapt==1.11.1
